Allow invalid SSL certificates in sslmode=prefer
Anyone can do a downgrade attack anyway, so there isn't really a point
in complaining about invalid certificates.

Setting sslmode=prefer is now the only way to disable cert validation:
I can't image a scenario where the network can be trusted to not
intercept TLS, but cannot be trusted to not do a downgrade attack.
Yes, downgrading is simpler, but if you care about that, just provide
the root_cert option as well.